requirements.txt
aiohttp           
aiosignal         
aniso8601         
appdirs           
asgiref           
async-timeout     
attrs              
beautifulsoup4     
bs4               
certifi           
charset-normalizer
click             
colorama          
cssselect         
fake-useragent     
Flask             
Flask-MySQLdb     
Flask-RESTful     
frozenlist        
idna            
importlib-metadata
itsdangerous      
Jinja2            
lxml              
MarkupSafe        
multidict         
mysqlclient  
mysql-connector-python     
nest-asyncio      
parse              
pip                
pyee              
pyppeteer         
pyquery           
python-decouple 
pytz               
requests           
requests-html      
setuptools         
six                
soupsieve         
tqdm               
typing_extensions 
urllib3            
w3lib              
websockets      
Werkzeug          
yarl              
zipp              